requirements.txt
Flask
python-dotenv
huggingface_hub